+++ packages/network-controller/tests/NetworkController.test.ts
@@ -6,6 +6,7 @@ import { ControllerMessenger } from '@metamask/base-controller';
 import * as ethQueryModule from 'eth-query';
 import { Patch } from 'immer';
 import { v4 } from 'uuid';
+import { ethErrors } from 'eth-rpc-errors';
 import {
   NetworkType,
   NetworksChainId,
@@ -26,6 +27,7 @@ import {
   NetworkState,
   ProviderConfig,
 } from '../src/NetworkController';
+import { NetworkStatus } from '../src/constants';
 import { BUILT_IN_NETWORKS } from '../../controller-utils/src/constants';
 import {
   createNetworkClient,
@@ -139,8 +141,8 @@ describe('NetworkController', () => {
       await withController(({ controller }) => {
         expect(controller.state).toStrictEqual({
           networkConfigurations: {},
-          network: 'loading',
-          isCustomNetwork: false,
+          networkId: null,
+          networkStatus: NetworkStatus.Unknown,
           providerConfig: { type: NetworkType.mainnet, chainId: '1' },
           networkDetails: { isEIP1559Compatible: false },
         });
@@ -151,15 +153,14 @@ describe('NetworkController', () => {
       await withController(
         {
           state: {
-            isCustomNetwork: true,
             networkDetails: { isEIP1559Compatible: true },
           },
         },
         ({ controller }) => {
           expect(controller.state).toStrictEqual({
             networkConfigurations: {},
-            network: 'loading',
-            isCustomNetwork: true,
+            networkId: null,
+            networkStatus: NetworkStatus.Unknown,
             providerConfig: { type: NetworkType.mainnet, chainId: '1' },
             networkDetails: { isEIP1559Compatible: true },
           });
@@ -169,55 +170,6 @@ describe('NetworkController', () => {
   });
 
   describe('initializeProvider', () => {
-    ['1', '5', '11155111', ''].forEach((chainId) => {
-      describe(`when the provider config in state contains a chain ID of "${chainId}"`, () => {
-        it('sets isCustomNetwork in state to false', async () => {
-          await withController(
-            {
-              state: {
-                isCustomNetwork: true,
-                providerConfig: buildProviderConfig({
-                  chainId,
-                }),
-              },
-              infuraProjectId: 'infura-project-id',
-            },
-            async ({ controller }) => {
-              const fakeJsonRpcClient = buildFakeClient();
-              createNetworkClientMock.mockReturnValue(fakeJsonRpcClient);
-
-              await controller.initializeProvider();
-
-              expect(controller.state.isCustomNetwork).toBe(false);
-            },
-          );
-        });
-      });
-    });
-
-    describe('when the provider config in state contains a chain ID that is not 1, 5, 11155111, or an empty string', () => {
-      it('sets isCustomNetwork in state to true', async () => {
-        await withController(
-          {
-            state: {
-              providerConfig: buildProviderConfig({
-                chainId: '999',
-              }),
-            },
-            infuraProjectId: 'infura-project-id',
-          },
-          async ({ controller }) => {
-            const fakeClient = buildFakeClient();
-            createNetworkClientMock.mockReturnValue(fakeClient);
-
-            await controller.initializeProvider();
-
-            expect(controller.state.isCustomNetwork).toBe(true);
-          },
-        );
-      });
-    });
-
     [NetworkType.mainnet, NetworkType.goerli, NetworkType.sepolia].forEach(
       (networkType) => {
         describe(`when the provider config in state contains a network type of "${networkType}"`, () => {
@@ -442,7 +394,7 @@ describe('NetworkController', () => {
         const messenger = buildMessenger();
         await withController({ messenger }, async ({ controller }) => {
           const promiseForNetworkChanges = waitForStateChanges(messenger, {
-            propertyPath: ['network'],
+            propertyPath: ['networkId'],
           });
 
           await controller.lookupNetwork();
@@ -471,7 +423,7 @@ describe('NetworkController', () => {
         it('updates the network in state to match', async () => {
           const messenger = buildMessenger();
           await withController(
-            { messenger, state: { network: '' } },
+            { messenger, state: { networkId: null } },
             async ({ controller }) => {
               await setFakeProvider(controller, {
                 stubs: [
@@ -484,13 +436,13 @@ describe('NetworkController', () => {
               });
 
               await waitForStateChanges(messenger, {
-                propertyPath: ['network'],
+                propertyPath: ['networkId'],
                 produceStateChanges: async () => {
                   await controller.lookupNetwork();
                 },
               });
 
-              expect(controller.state.network).toBe('12345');
+              expect(controller.state.networkId).toBe('12345');
             },
           );
         });
@@ -545,7 +497,7 @@ describe('NetworkController', () => {
         it('does not change network in state', async () => {
           const messenger = buildMessenger();
           await withController(
-            { messenger, state: { network: '12345' } },
+            { messenger, state: { networkId: '12345' } },
             async ({ controller }) => {
               await setFakeProvider(controller, {
                 stubs: [
@@ -557,7 +509,7 @@ describe('NetworkController', () => {
                 stubLookupNetworkWhileSetting: true,
               });
               const promiseForNetworkChanges = waitForStateChanges(messenger, {
-                propertyPath: ['network'],
+                propertyPath: ['networkId'],
               });
 
               await controller.lookupNetwork();
@@ -570,7 +522,7 @@ describe('NetworkController', () => {
         it('does not publish NetworkController:providerConfigChange', async () => {
           const messenger = buildMessenger();
           await withController(
-            { messenger, state: { network: '12345' } },
+            { messenger, state: { networkId: '12345' } },
             async ({ controller }) => {
               await setFakeProvider(controller, {
                 stubs: [
@@ -594,30 +546,32 @@ describe('NetworkController', () => {
         });
       });
 
-      describe('if an error is encountered while retrieving the version of the current network', () => {
-        it('updates the network in state to "loading"', async () => {
+      describe('if an RPC error is encountered while retrieving the version of the current network', () => {
+        it('updates the network in state to "unavailable"', async () => {
           const messenger = buildMessenger();
           await withController(
-            { messenger, state: { network: '1' } },
+            { messenger, state: { networkId: '1' } },
             async ({ controller }) => {
               await setFakeProvider(controller, {
                 stubs: [
                   {
                     request: { method: 'net_version' },
-                    response: { error: 'some error' },
+                    error: ethErrors.rpc.limitExceeded('some error'),
                   },
                 ],
                 stubLookupNetworkWhileSetting: true,
               });
 
               await waitForStateChanges(messenger, {
-                propertyPath: ['network'],
+                propertyPath: ['networkId'],
                 produceStateChanges: async () => {
                   await controller.lookupNetwork();
                 },
               });
 
-              expect(controller.state.network).toBe('loading');
+              expect(controller.state.networkStatus).toBe(
+                NetworkStatus.Unavailable,
+              );
             },
           );
         });
@@ -639,7 +593,159 @@ describe('NetworkController', () => {
                 stubs: [
                   {
                     request: { method: 'net_version' },
-                    response: { error: 'some error' },
+                    error: ethErrors.rpc.limitExceeded('some error'),
+                  },
+                ],
+                stubLookupNetworkWhileSetting: true,
+              });
+
+              const providerConfigChanges = await waitForPublishedEvents(
+                messenger,
+                'NetworkController:providerConfigChange',
+                {
+                  produceEvents: async () => {
+                    await controller.lookupNetwork();
+                  },
+                },
+              );
+
+              expect(providerConfigChanges).toStrictEqual([
+                [
+                  {
+                    type: NetworkType.mainnet,
+                    chainId: '1',
+                  },
+                ],
+              ]);
+            },
+          );
+        });
+      });
+
+      describe('if an internal error is encountered while retrieving the version of the current network', () => {
+        it('updates the network in state to "unknown"', async () => {
+          const messenger = buildMessenger();
+          await withController(
+            { messenger, state: { networkId: '1' } },
+            async ({ controller }) => {
+              await setFakeProvider(controller, {
+                stubs: [
+                  {
+                    request: { method: 'net_version' },
+                    error: ethErrors.rpc.internal('some error'),
+                  },
+                ],
+                stubLookupNetworkWhileSetting: true,
+              });
+
+              await waitForStateChanges(messenger, {
+                propertyPath: ['networkId'],
+                produceStateChanges: async () => {
+                  await controller.lookupNetwork();
+                },
+              });
+
+              expect(controller.state.networkStatus).toBe(
+                NetworkStatus.Unknown,
+              );
+            },
+          );
+        });
+
+        it("publishes NetworkController:providerConfigChange with the current provider config (even though it didn't change)", async () => {
+          const messenger = buildMessenger();
+          await withController(
+            {
+              messenger,
+              state: {
+                providerConfig: {
+                  type: NetworkType.mainnet,
+                  chainId: '1',
+                },
+              },
+            },
+            async ({ controller }) => {
+              await setFakeProvider(controller, {
+                stubs: [
+                  {
+                    request: { method: 'net_version' },
+                    error: ethErrors.rpc.internal('some error'),
+                  },
+                ],
+                stubLookupNetworkWhileSetting: true,
+              });
+
+              const providerConfigChanges = await waitForPublishedEvents(
+                messenger,
+                'NetworkController:providerConfigChange',
+                {
+                  produceEvents: async () => {
+                    await controller.lookupNetwork();
+                  },
+                },
+              );
+
+              expect(providerConfigChanges).toStrictEqual([
+                [
+                  {
+                    type: NetworkType.mainnet,
+                    chainId: '1',
+                  },
+                ],
+              ]);
+            },
+          );
+        });
+      });
+
+      describe('if an invalid network ID is returned', () => {
+        it('updates the network in state to "unknown"', async () => {
+          const messenger = buildMessenger();
+          await withController(
+            { messenger, state: { networkId: '1' } },
+            async ({ controller }) => {
+              await setFakeProvider(controller, {
+                stubs: [
+                  {
+                    request: { method: 'net_version' },
+                    response: { result: 'invalid' },
+                  },
+                ],
+                stubLookupNetworkWhileSetting: true,
+              });
+
+              await waitForStateChanges(messenger, {
+                propertyPath: ['networkId'],
+                produceStateChanges: async () => {
+                  await controller.lookupNetwork();
+                },
+              });
+
+              expect(controller.state.networkStatus).toBe(
+                NetworkStatus.Unknown,
+              );
+            },
+          );
+        });
+
+        it("publishes NetworkController:providerConfigChange with the current provider config (even though it didn't change)", async () => {
+          const messenger = buildMessenger();
+          await withController(
+            {
+              messenger,
+              state: {
+                providerConfig: {
+                  type: NetworkType.mainnet,
+                  chainId: '1',
+                },
+              },
+            },
+            async ({ controller }) => {
+              await setFakeProvider(controller, {
+                stubs: [
+                  {
+                    request: { method: 'net_version' },
+                    response: { result: 'invalid' },
                   },
                 ],
                 stubLookupNetworkWhileSetting: true,
@@ -693,7 +799,7 @@ describe('NetworkController', () => {
               stubLookupNetworkWhileSetting: true,
             });
             const promiseForStateChanges = waitForStateChanges(messenger, {
-              propertyPath: ['network'],
+              propertyPath: ['networkId'],
               count: 3,
               wait: 400,
             });
@@ -706,13 +812,13 @@ describe('NetworkController', () => {
 
             expect(await promiseForStateChanges).toMatchObject([
               expect.objectContaining([
-                expect.objectContaining({ network: '1' }),
+                expect.objectContaining({ networkId: '1' }),
               ]),
               expect.objectContaining([
-                expect.objectContaining({ network: '2' }),
+                expect.objectContaining({ networkId: '2' }),
               ]),
               expect.objectContaining([
-                expect.objectContaining({ network: '3' }),
+                expect.objectContaining({ networkId: '3' }),
               ]),
             ]);
           });
@@ -722,156 +828,13 @@ describe('NetworkController', () => {
   });
 
   describe('setProviderType', () => {
-    describe('given a network type of "mainnet"', () => {
-      it('updates the provider config in state with the network type and the corresponding chain ID, using "ETH" for the ticker and clearing any existing RPC target and nickname', async () => {
-        const messenger = buildMessenger();
-        await withController(
-          {
-            messenger,
-            state: {
-              providerConfig: {
-                type: NetworkType.localhost,
-                rpcTarget: 'http://somethingexisting.com',
-                chainId: '99999',
-                ticker: 'something existing',
-                nickname: 'something existing',
-              },
-            },
-          },
-          async ({ controller }) => {
-            const fakeInfuraClient = buildFakeClient();
-            createNetworkClientMock.mockReturnValue(fakeInfuraClient);
-
-            await controller.setProviderType(NetworkType.mainnet);
-
-            expect(controller.state.providerConfig).toStrictEqual({
-              type: NetworkType.mainnet,
-              ...BUILT_IN_NETWORKS.mainnet,
-              rpcTarget: undefined,
-              nickname: undefined,
-              id: undefined,
-            });
-          },
-        );
-      });
-
-      it('sets isCustomNetwork in state to false', async () => {
-        const messenger = buildMessenger();
-        await withController(
-          {
-            messenger,
-            state: {
-              isCustomNetwork: true,
-            },
-            infuraProjectId: 'infura-project-id',
-          },
-          async ({ controller }) => {
-            const fakeInfuraClient = buildFakeClient();
-            createNetworkClientMock.mockReturnValue(fakeInfuraClient);
-
-            await controller.setProviderType(NetworkType.mainnet);
-
-            expect(controller.state.isCustomNetwork).toBe(false);
-          },
-        );
-      });
-
-      it('sets the provider to an Infura provider pointed to Mainnet', async () => {
-        await withController(
-          {
-            infuraProjectId: 'infura-project-id',
-          },
-          async ({ controller }) => {
-            const fakeProvider = buildFakeProvider([
-              {
-                request: {
-                  method: 'eth_chainId',
-                },
-                response: {
-                  result: '0x1337',
-                },
-              },
-            ]);
-            const fakeInfuraClient = buildFakeClient(fakeProvider);
-            createNetworkClientMock.mockReturnValue(fakeInfuraClient);
-
-            await controller.setProviderType(NetworkType.mainnet);
-
-            expect(createNetworkClientMock).toHaveBeenCalledWith({
-              network: NetworkType.mainnet,
-              infuraProjectId: 'infura-project-id',
-              type: NetworkClientType.Infura,
-            });
-            const { provider } = controller.getProviderAndBlockTracker();
-            const promisifiedSendAsync = promisify(provider.sendAsync).bind(
-              provider,
-            );
-            const chainIdResult = await promisifiedSendAsync({
-              method: 'eth_chainId',
-            });
-            expect(chainIdResult.result).toBe('0x1337');
-          },
-        );
-      });
-
-      it('updates networkDetails.isEIP1559Compatible in state based on the latest block (assuming that the request for eth_getBlockByNumber is made successfully)', async () => {
-        const messenger = buildMessenger();
-        await withController(
-          {
-            messenger,
-          },
-          async ({ controller }) => {
-            const fakeProvider = buildFakeProvider([
-              {
-                request: {
-                  method: 'eth_getBlockByNumber',
-                  params: ['latest', false],
-                },
-                response: {
-                  result: {
-                    baseFeePerGas: '0x1',
-                  },
-                },
-              },
-            ]);
-            const fakeInfuraClient = buildFakeClient(fakeProvider);
-            createNetworkClientMock.mockReturnValue(fakeInfuraClient);
-
-            await controller.setProviderType(NetworkType.mainnet);
-
-            expect(controller.state.networkDetails.isEIP1559Compatible).toBe(
-              true,
-            );
-          },
-        );
-      });
-
-      it('records the version of the current network in state (assuming that the request for net_version is made successfully)', async () => {
-        const messenger = buildMessenger();
-        await withController({ messenger }, async ({ controller }) => {
-          const fakeProvider = buildFakeProvider([
-            {
-              request: {
-                method: 'net_version',
-                params: [],
-              },
-              response: {
-                result: '42',
-              },
-            },
-          ]);
-          const fakeInfuraClient = buildFakeClient(fakeProvider);
-          createNetworkClientMock.mockReturnValue(fakeInfuraClient);
-
-          await controller.setProviderType(NetworkType.mainnet);
-
-          expect(controller.state.network).toBe('42');
-        });
-      });
-    });
-
     (
       [
+        {
+          networkType: NetworkType.mainnet,
+          ticker: NetworksTicker.mainnet,
+          chainId: NetworksChainId.mainnet,
+        },
         {
           networkType: NetworkType.goerli,
           ticker: NetworksTicker.goerli,
@@ -913,26 +876,6 @@ describe('NetworkController', () => {
                 nickname: undefined,
                 id: undefined,
               });
-            },
-          );
-        });
-
-        it('sets isCustomNetwork in state to false', async () => {
-          const messenger = buildMessenger();
-          await withController(
-            {
-              messenger,
-              state: {
-                isCustomNetwork: true,
-              },
-            },
-            async ({ controller }) => {
-              const fakeInfuraClient = buildFakeClient();
-              createNetworkClientMock.mockReturnValue(fakeInfuraClient);
-
-              await controller.setProviderType(networkType);
-
-              expect(controller.state.isCustomNetwork).toBe(false);
             },
           );
         });
@@ -1092,26 +1035,6 @@ describe('NetworkController', () => {
         );
       });
 
-      it('sets isCustomNetwork in state to false', async () => {
-        const messenger = buildMessenger();
-        await withController(
-          {
-            messenger,
-            state: {
-              isCustomNetwork: true,
-            },
-          },
-          async ({ controller }) => {
-            const fakeJsonRpcClient = buildFakeClient();
-            createNetworkClientMock.mockReturnValue(fakeJsonRpcClient);
-
-            await controller.setProviderType(NetworkType.localhost);
-
-            expect(controller.state.isCustomNetwork).toBe(false);
-          },
-        );
-      });
-
       it('sets the provider to a custom RPC provider pointed to localhost, leaving chain ID undefined', async () => {
         await withController(async ({ controller }) => {
           const fakeProvider = buildFakeProvider([
@@ -1191,7 +1114,7 @@ describe('NetworkController', () => {
 
           await controller.setProviderType(NetworkType.localhost);
 
-          expect(controller.state.network).toBe('42');
+          expect(controller.state.networkId).toBe('42');
         });
       });
     });
@@ -1246,36 +1169,6 @@ describe('NetworkController', () => {
             nickname: undefined,
             rpcPrefs: undefined,
           });
-        },
-      );
-    });
-
-    it('sets isCustomNetwork in state to true', async () => {
-      const messenger = buildMessenger();
-      await withController(
-        {
-          messenger,
-          state: {
-            isCustomNetwork: false,
-            networkConfigurations: {
-              testNetworkConfigurationId: {
-                rpcUrl: 'https://mock-rpc-url',
-                chainId: '0xabc',
-                ticker: 'TEST',
-                id: 'testNetworkConfigurationId',
-                nickname: undefined,
-                rpcPrefs: undefined,
-              },
-            },
-          },
-        },
-        async ({ controller }) => {
-          const fakeClient = buildFakeClient();
-          createNetworkClientMock.mockReturnValue(fakeClient);
-
-          await controller.setActiveNetwork('testNetworkConfigurationId');
-
-          expect(controller.state.isCustomNetwork).toBe(true);
         },
       );
     });
@@ -1337,7 +1230,6 @@ describe('NetworkController', () => {
         {
           messenger,
           state: {
-            isCustomNetwork: false,
             networkConfigurations: {
               testNetworkConfigurationId: {
                 rpcUrl: 'https://mock-rpc-url',
@@ -1411,7 +1303,7 @@ describe('NetworkController', () => {
 
           await controller.setActiveNetwork('testNetworkConfigurationId');
 
-          expect(controller.state.network).toBe('42');
+          expect(controller.state.networkId).toBe('42');
         },
       );
     });
@@ -2200,32 +2092,6 @@ describe('NetworkController', () => {
     [NetworkType.mainnet, NetworkType.goerli, NetworkType.sepolia].forEach(
       (networkType) => {
         describe(`when the type in the provider configuration is "${networkType}"`, () => {
-          it('sets isCustomNetwork in state to false', async () => {
-            const messenger = buildMessenger();
-            await withController(
-              {
-                messenger,
-                state: {
-                  isCustomNetwork: true,
-                },
-                infuraProjectId: 'infura-project-id',
-              },
-              async ({ controller }) => {
-                const fakeClient = buildFakeClient();
-                createNetworkClientMock.mockReturnValue(fakeClient);
-
-                await waitForStateChanges(messenger, {
-                  propertyPath: ['isCustomNetwork'],
-                  produceStateChanges: () => {
-                    controller.resetConnection();
-                  },
-                });
-
-                expect(controller.state.isCustomNetwork).toBe(false);
-              },
-            );
-          });
-
           it(`initializes a new provider object pointed to the current Infura network (type: "${networkType}")`, async () => {
             await withController(
               {
@@ -2340,39 +2206,6 @@ describe('NetworkController', () => {
     );
 
     describe(`when the type in the provider configuration is "rpc"`, () => {
-      it('sets isCustomNetwork in state to true', async () => {
-        const messenger = buildMessenger();
-        await withController(
-          {
-            messenger,
-            state: {
-              isCustomNetwork: false,
-              providerConfig: {
-                type: NetworkType.rpc,
-                rpcTarget: 'https://mock-rpc-url',
-                chainId: '0x1337',
-                ticker: 'TEST',
-                id: 'testNetworkConfigurationId',
-              },
-            },
-            infuraProjectId: 'infura-project-id',
-          },
-          async ({ controller }) => {
-            const fakeClient = buildFakeClient();
-            createNetworkClientMock.mockReturnValue(fakeClient);
-
-            await waitForStateChanges(messenger, {
-              propertyPath: ['isCustomNetwork'],
-              produceStateChanges: () => {
-                controller.resetConnection();
-              },
-            });
-
-            expect(controller.state.isCustomNetwork).toBe(true);
-          },
-        );
-      });
-
       it('initializes a new provider object pointed to the same RPC URL as the current network and using the same chain ID', async () => {
         await withController(
           {
@@ -3413,10 +3246,10 @@ describe('NetworkController', () => {
           );
         });
 
-        it('resets isEIP1559Compatible and sets network to "loading" for the network before emitting NetworkController:providerConfigChange', async () => {
+        it('resets isEIP1559Compatible and sets network to "unknown" for the network before emitting NetworkController:providerConfigChange', async () => {
           const networkConfiguration = {
             rpcUrl: 'https://mock-rpc-url',
-            chainId: '0xabc',
+            chainId: '0x1338',
             ticker: 'TEST',
             nickname: undefined,
             id: 'testNetworkConfigurationId',
@@ -3452,6 +3285,15 @@ describe('NetworkController', () => {
                   },
                 },
                 {
+                  delay: 1,
+                  request: {
+                    method: 'net_version',
+                  },
+                  response: {
+                    result: '1338',
+                  },
+                },
+                {
                   request: {
                     method: 'eth_getBlockByNumber',
                   },
@@ -3460,11 +3302,12 @@ describe('NetworkController', () => {
                   },
                 },
                 {
+                  delay: 1,
                   request: {
                     method: 'net_version',
                   },
                   response: {
-                    result: '10000',
+                    result: '1337',
                   },
                 },
               ]);
@@ -3476,13 +3319,15 @@ describe('NetworkController', () => {
                 isEIP1559Compatible: true,
               });
               await waitForStateChanges(messenger, {
-                propertyPath: ['network'],
+                propertyPath: ['networkId'],
                 count: 1,
                 produceStateChanges: () => {
                   controller.rollbackToPreviousProvider();
                 },
               });
-              expect(controller.state.network).toStrictEqual('loading');
+              expect(controller.state.networkStatus).toBe(
+                NetworkStatus.Unknown,
+              );
               expect(controller.state.networkDetails).toStrictEqual({
                 isEIP1559Compatible: false,
               });
@@ -3535,7 +3380,7 @@ describe('NetworkController', () => {
 
               await controller.setActiveNetwork('testNetworkConfigurationId');
               await waitForStateChanges(messenger, {
-                propertyPath: ['network'],
+                propertyPath: ['networkId'],
                 count: 1,
                 produceStateChanges: () => {
                   controller.rollbackToPreviousProvider();
@@ -3589,7 +3434,7 @@ describe('NetworkController', () => {
                 controller.getProviderAndBlockTracker();
 
               await waitForStateChanges(messenger, {
-                propertyPath: ['network'],
+                propertyPath: ['networkId'],
                 count: 1,
                 produceStateChanges: () => {
                   controller.rollbackToPreviousProvider();
@@ -3654,16 +3499,16 @@ describe('NetworkController', () => {
               createNetworkClientMock.mockReturnValue(fakeJsonRpcClient);
 
               await controller.setActiveNetwork('testNetworkConfigurationId');
-              expect(controller.state.network).toStrictEqual('999');
+              expect(controller.state.networkId).toStrictEqual('999');
 
               await waitForStateChanges(messenger, {
-                propertyPath: ['network'],
+                propertyPath: ['networkId'],
                 count: 1,
                 produceStateChanges: () => {
                   controller.rollbackToPreviousProvider();
                 },
               });
-              expect(controller.state.network).toStrictEqual('1');
+              expect(controller.state.networkId).toStrictEqual('1');
             },
           );
         });
@@ -3838,7 +3683,7 @@ describe('NetworkController', () => {
         );
       });
 
-      it('resets the network state to "loading" and the isEIP before emitting NetworkController:providerConfigChange', async () => {
+      it('resets the network state to "unknown" and the isEIP before emitting NetworkController:providerConfigChange', async () => {
         const messenger = buildMessenger();
         const initialProviderConfigNetworkConfiguration = {
           rpcUrl: 'https://mock-rpc-url-2',
@@ -3862,7 +3707,7 @@ describe('NetworkController', () => {
               networkConfigurations: {
                 testNetworkConfigurationId1: {
                   rpcUrl: 'https://mock-rpc-url',
-                  chainId: '0xabc',
+                  chainId: '0x1338',
                   ticker: 'TEST',
                   id: 'testNetworkConfigurationId1',
                 },
@@ -3882,6 +3727,15 @@ describe('NetworkController', () => {
                 },
               },
               {
+                delay: 1,
+                request: {
+                  method: 'net_version',
+                },
+                response: {
+                  result: '1338',
+                },
+              },
+              {
                 request: {
                   method: 'eth_getBlockByNumber',
                 },
@@ -3890,11 +3744,12 @@ describe('NetworkController', () => {
                 },
               },
               {
+                delay: 1,
                 request: {
                   method: 'net_version',
                 },
                 response: {
-                  result: '10000',
+                  result: '1337',
                 },
               },
             ]);
@@ -3904,14 +3759,19 @@ describe('NetworkController', () => {
             expect(controller.state.networkDetails).toStrictEqual({
               isEIP1559Compatible: true,
             });
+            expect(controller.state.networkStatus).toStrictEqual(
+              NetworkStatus.Available,
+            );
             await waitForStateChanges(messenger, {
-              propertyPath: ['network'],
+              propertyPath: ['networkStatus'],
               count: 1,
               produceStateChanges: () => {
                 controller.rollbackToPreviousProvider();
               },
             });
-            expect(controller.state.network).toStrictEqual('loading');
+            expect(controller.state.networkStatus).toStrictEqual(
+              NetworkStatus.Unknown,
+            );
             expect(controller.state.networkDetails).toStrictEqual({
               isEIP1559Compatible: false,
             });
@@ -3972,7 +3832,7 @@ describe('NetworkController', () => {
 
             await controller.setActiveNetwork('testNetworkConfigurationId1');
             await waitForStateChanges(messenger, {
-              propertyPath: ['network'],
+              propertyPath: ['networkId'],
               count: 1,
               produceStateChanges: () => {
                 controller.rollbackToPreviousProvider();
@@ -4037,7 +3897,7 @@ describe('NetworkController', () => {
               controller.getProviderAndBlockTracker();
 
             await waitForStateChanges(messenger, {
-              propertyPath: ['network'],
+              propertyPath: ['networkId'],
               count: 1,
               produceStateChanges: () => {
                 controller.rollbackToPreviousProvider();
@@ -4111,16 +3971,16 @@ describe('NetworkController', () => {
             createNetworkClientMock.mockReturnValue(fakeJsonRpcClient);
 
             await controller.setActiveNetwork('testNetworkConfigurationId1');
-            expect(controller.state.network).toStrictEqual('999');
+            expect(controller.state.networkId).toStrictEqual('999');
 
             await waitForStateChanges(messenger, {
-              propertyPath: ['network'],
+              propertyPath: ['networkId'],
               count: 1,
               produceStateChanges: () => {
                 controller.rollbackToPreviousProvider();
               },
             });
-            expect(controller.state.network).toStrictEqual('1');
+            expect(controller.state.networkId).toStrictEqual('1');
           },
         );
       });